+++ packages/expect/src/__tests__/matchers.test.js
@@ -435,6 +435,17 @@ describe('.toEqual()', () => {
     [{a: 1}, {a: 2}],
     [{a: 5}, {b: 6}],
     [Object.freeze({foo: {bar: 1}}), {foo: {}}],
+    [
+      {
+        get a() {
+          return 'foo';
+        },
+        set a(value) {
+          throw new Error('noo');
+        },
+      },
+      {a: 'bar'},
+    ],
     ['banana', 'apple'],
     ['1\u{00A0}234,57\u{00A0}$', '1 234,57 $'], // issues/6881
     [