+++ packages/jest-transform/src/index.ts
@@ -11,6 +11,7 @@ export {
 } from './ScriptTransformer';
 export {default as shouldInstrument} from './shouldInstrument';
 export type {
+  CacheKeyOptions,
   Transformer,
   ShouldInstrumentOptions,
   Options as TransformationOptions,