+++ CHANGELOG.md
@@ -3,8 +3,7 @@
 ### Features
 
 ### Fixes
-
-- `[@jest/watcher]` Correct return type of `shouldRunTestSuite` for `JestHookEmitter`  ([#9753](https://github.com/facebook/jest/pull/9753))
+- `[@jest/watcher]` Correct return type of `shouldRunTestSuite` for `JestHookEmitter` ([#9753](https://github.com/facebook/jest/pull/9753))
 
 ### Chore & Maintenance