+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Features
 
 ### Fixes
+
+- `[jest-matcher-utils]` Do not override properties with setters when diffing objects
 - `[@jest/watcher]` Correct return type of `shouldRunTestSuite` for `JestHookEmitter` ([#9753](https://github.com/facebook/jest/pull/9753))
 
 ### Chore & Maintenance