+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ### Chore & Maintenance
 
-- `[@jest/transform]` Expose type `TransformedSource` ([#9762](https://github.com/facebook/jest/pull/9762))
+- `[@jest/transform]` Expose type `CacheKeyOptions` for `getCacheKey` ([#9762](https://github.com/facebook/jest/pull/9762))
 
 ### Performance