+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Features
 
-- `[babel-jest]` Support passing `supportsDynamicImport` and `supportsStaticESM`
+- `[babel-jest]` Support passing `supportsDynamicImport` and `supportsStaticESM` ([#9766](https://github.com/facebook/jest/pull/9766))
 
 ### Fixes