+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Features
 
+- `[babel-jest]` Support passing `supportsDynamicImport` and `supportsStaticESM` ([#9766](https://github.com/facebook/jest/pull/9766))
+
 ### Fixes
 
 ### Chore & Maintenance